docs: document new route-specific fields
Move InitialView, MagnificationOption, and PageLayout enums to
gotenberg_client.options (public API) and update CHANGELOG.md,
docs/routes.md, and docs/ref_options.md to cover all new fields
added in Tasks 1–6.
